User can add one todo item
Which exactly corresponds to whatever they have typed in the input,
and which is not persisted. However, it does pass the acceptance test!
Obviously, the AT is not very good since the app is not useful, but how
specific should the AT be? I want to do ATDD, but I'm not sure how.

A todo app should be an interesting case study because there are so many
of them and a plain text file opened in Notepad could perfectly satisfy
the requirements I have stated so far.

Anyway, a specific question to research: Given that I have a high-level
description of the acceptance criteria, when do I implement the details
of the queries and commands to integrate the test with the app? In this
case, I had to specify up front that there would be an element on the
page with the id 'new-item' into which the user can type a todo item
that will then show up somewhere else on the page. Also, *there has to
be a page!* By using Cypress to write the AT, I'm committing myself to
the choice that this is a web app. The ATs will be inextricably coupled
to the web.