Extend JWTAuthorizationFilter from simpler base class
The Auth0 blog post about implementing JWT auth for Spring Boot[0]
that I adapted my implemention from used BasicAuthenticationFilter, so
that's what I did, too. It seemed weird to me, so I poked around and it
doesn't seem to be necessary. Extending OncePerRequestFilter (which is
the base class for BasicAuthenticationFilter) works just fine.

The only problem was that Spring started complaining that it didn't know
what order to run the JWTAuthorizationFilter in, hence the change to
SecurityConfiguration. I somewhat arbitrarily chose to make it run after
BasicAuthenticationFilter.

0: https://auth0.com/blog/implementing-jwt-authentication-on-spring-boot